requirements.txt
transformers
torch
pillow
pytesseract
SpeechRecognition
pyaudio
scikit-learn
datasets
pytorch-lightning
gradio
